requirements.txt
pydantic>=2.9.0
pyautogen[openai]==0.2.21
